Optimizations from PR feedback
